% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{check_dataframe_validity}
\alias{check_dataframe_validity}
\title{Validate a Data Frame for LCMD Imputation}
\usage{
check_dataframe_validity(df)
}
\arguments{
\item{df}{An object to be validated.}
}
\value{
TRUE if the input passes all checks; otherwise throws an error.
}
\description{
Ensures the input is a data frame with at least 1 column, 2 rows,
and that all columns are numeric.
}
