% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{data_check}
\alias{data_check}
\title{check required columns}
\usage{
data_check(
  data,
  data_meta_features,
  data_meta_samples,
  col_samples,
  col_features = NULL,
  exclusion_extreme_feature = FALSE,
  exclusion_extreme_sample = FALSE,
  missing_pct_feature = NULL,
  missing_pct_sample = NULL
)
}
\arguments{
\item{data}{Data frame containing sample data, which must include \code{col_samples}.}

\item{data_meta_features}{Data frame containing feature metadata, which must include
\code{col_features}}

\item{data_meta_samples}{Data frame containing sample metadata, which must include
\code{col_samples}}

\item{col_samples}{Character vector specifying the column name(s) in \code{data} representing sample IDs.}

\item{col_features}{Character string specifying the column name in \code{data_meta_features} representing
feature IDs (optional, only checked if provided).}

\item{exclusion_extreme_feature}{Logical; if \code{TRUE}, \code{missing_pct_feature}
must be provided and exist in \code{data_meta_features}.}

\item{exclusion_extreme_sample}{Logical; if \code{TRUE}, \code{missing_pct_sample}
must be provided and exist in \code{data_meta_samples}.}

\item{missing_pct_feature}{Numeric value indicating the acceptable percentage of missing data for features
(required if \code{exclusion_extreme_feature} is \code{TRUE}).}

\item{missing_pct_sample}{Numeric value indicating the acceptable percentage of missing data for samples
(required if \code{exclusion_extreme_sample} is \code{TRUE}).}
}
\value{
Logical; \code{TRUE} if all required columns are present, otherwise an error is raised.
}
\description{
This function checks for the presence of required columns
It verifies that necessary columns are present based on specified exclusion flags.
}
