% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_features}
\alias{data_features}
\title{Simulated Dataframe 1}
\format{
A data frame with 100 rows and 101 variables:
}
\usage{
data(data_features)
}
\description{
A simulated dataframe with 100 rows and 101 columns. Col1 is an ID column.
The remaining columns are features with randomly assigned values. Once values
were assigned, missingness was introduced (5 percent) randomly for feature_2-feature92.
feature_1 has no missing; feature_93-feature_100 contains missing at 10, 20,
30, 40, 50, 60, 70, 80, and 90 perceent, respectively.
}
\keyword{datasets}
