% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_meta_features}
\alias{data_meta_features}
\title{Simulated Dataframe 2}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{ID_feature}{Feature names from data}
\item{missing_pct}{percentage of missing values for each feature}
\item{LOD}{limit of detection is the minimum feature value / 2}
}
}
\usage{
data(data_meta_features)
}
\description{
A dataframe summarizing missing values and minimum values from df1.
}
\keyword{datasets}
