% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_meta_samples}
\alias{data_meta_samples}
\title{Simulated Dataframe 3}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{ID_sample}{ID from ID_sample column in data}
\item{batch}{randomly assigned levels (N = 10)}
\item{sex}{randomly assigned 1 or 2}
\item{age}{mean 55; SD 10}
}
}
\usage{
data(data_meta_samples)
}
\description{
A dataframe with ID, batch, sex, and age information.
}
\keyword{datasets}
