% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_registration.R
\name{unregister_cluster}
\alias{unregister_cluster}
\title{Unregister a parallel cluster used for parallel processing.}
\usage{
unregister_cluster(cluster = NULL)
}
\arguments{
\item{cluster}{The parallel cluster object to be unregistered (default is NULL).}
}
\description{
This function unregisters a previously registered parallel cluster used for parallel processing in R.
}
