% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{exclude_features}
\alias{exclude_features}
\title{Exclude features with X misingness}
\usage{
exclude_features(df, df_features = NULL, missing_pct_feature = NULL)
}
\arguments{
\item{df}{A data frame of feature data only}

\item{df_features}{A data frame containing metadata for features}

\item{missing_pct_feature}{Numeric, the threshold (between 0 and 1) for feature missing data exclusion (e.g., 0.2 for 20\%).}
}
\value{
A list containing:
\item{df}{The filtered data frame with excluded features removed.}
\item{df_features}{The filtered feature metadata with excluded features removed.}
}
\description{
This function excludes features from a dataset (\code{df}) based on a specified
threshold of missing data percentage in a meta data file.
}
